Add auth/signup route testing
Create testing file for the route. The tests use superset to handle the
requests.
To avoid connecting to the database in testing environement, the NODE_ENV
variable is checked before the connection is done.

A babel plugin is now used to transpile import.meta, to make it usable
in jest.

Since nodemon is a development package, it has been switched from depen-
dencies to devDependencies.