Update multer filename parameter
Multer is giving a new name to the uploaded files. Before, it was simply
appending the date and the extension.

To avoid using characters that may not be handled on different OS, the new
filename is filtered and only contains lowercase and uppercase letters,
digits, dashes, underscores and periods.

The ancient extension was kept on the filename, resulting in things like
"imagename.png0000000.png". It is now removed.