Updates bodyJsonParse
The middleware was throwing an error if the property was undefined.

In some valid cases, the property may be undefined. For example, if the
user has the choice between sending an image or not. If he sends an ima-
ge, the property containing other data is defined. But if he doesn't,
it is undefined and valid.

The programmer now has the choice between throwing or not when creating
the middleware.