Better error logging
Use the errors formater to use Errors as logging messages.

Update the formating function to display the stacktrace.

Log uncaught exceptions and unhandled rejections as fatals. I didn't use
the built in winston functionnality to do so, because it was using the
error level and was not printing the stacktrace.